Adjust the code structure of keras.unittest.
