+++ test/TensorFlowNET.Keras.UnitTest/InitializerTest.cs
@@ -1,12 +1,7 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using TensorFlowNET.Keras.UnitTest;
 using static Tensorflow.Binding;
 
-namespace TensorFlowNET.Keras.UnitTest;
+namespace Tensorflow.Keras.UnitTest;
 
 [TestClass]
 public class InitializerTest : EagerModeTestBase
